backend/llm_service/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
google-generativeai
httpx
redis
python-dotenv
